Add files via upload
Arslan, a data analyst and Kaggle expert created data visualizations of the Pakistan Budget 2023-24. The visualizations provide insights into the budget's spending priorities, including its focus on infrastructure, social development, and economic growth. The visualizations are clear, concise, and easy to understand, making them a valuable resource for policymakers, businesses, and the general public. The visualizations are available on Arslan's GitHub repository, where they can be downloaded and reused.